fix(supabase): narrow init cli internals
